requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
python-socketio==5.10.0
eventlet==0.33.3
SQLAlchemy==2.0.21
Werkzeug==2.3.7
WTForms==3.0.1
alembic==1.12.0
email-validator==2.0.0.post2
python-dotenv==1.0.0
mysqlclient==2.2.0
xlsxwriter==3.1.9
Pillow==10.0.1 